# docs/swagger.yaml
basePath: /api/v1
definitions:
  bengkel_internal_http_response.APIError:
    properties:
      code:
        type: string
      details:
        items:
          $ref: '#/definitions/bengkel_internal_http_response.FieldError'
        type: array
      message:
        type: string
    type: object
  bengkel_internal_http_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/bengkel_internal_http_response.APIError'
      meta:
        $ref: '#/definitions/bengkel_internal_http_response.Meta'
    type: object
  bengkel_internal_http_response.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
      rule:
        type: string
    type: object
  bengkel_internal_http_response.Meta:
    properties:
      last_page:
        type: integer
      page:
        type: integer
      per_page:
        type: integer
      request_id:
        type: string
      total:
        type: integer
    type: object
...
  internal_http_handler.roleInput:
    properties:
      code:
        maxLength: 60
        type: string
      description:
        type: string
      name:
        maxLength: 100
        type: string
      permission_ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - code
    - name
    - permission_ids
    type: object
...
info:
  contact: {}
  description: API operasional bengkel, kasir, stok, pembayaran, akuntansi, audit,
    dan CMS.
  title: BengkelOS API
  version: "1.0"
paths:
...
  /permissions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Katalog permission
      tags:
      - Authorization
...
  /roles:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Daftar role beserta permission
      tags:
      - Authorization
    post:
      consumes:
      - application/json
      parameters:
      - description: Role
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_http_handler.roleInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Tambah role dan permission
      tags:
      - Authorization
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Masukkan "Bearer {access_token}".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"